+++ examples/blacklist/provider.tf
@@ -1,6 +1,6 @@
 
 terraform {
-  required_version = ">=0.12"
+  required_version = ">= 1.0"
 
   required_providers {
     guance = {